Move app_session into runtime/
