+++ lib/streamlit/web/server/server.py
@@ -42,7 +42,7 @@ from streamlit import config
 from streamlit import file_util
 from streamlit import source_util
 from streamlit import util
-from streamlit.app_session import AppSession
+from streamlit.runtime.app_session import AppSession
 from streamlit.caching import get_memo_stats_provider, get_singleton_stats_provider
 from streamlit.components.v1.components import ComponentRegistry
 from streamlit.config_option import ConfigOption
