+++ lib/tests/streamlit/app_session_test.py
@@ -20,9 +20,9 @@ from unittest.mock import MagicMock, patch
 
 import pytest
 
-import streamlit.app_session as app_session
+import streamlit.runtime.app_session as app_session
 from streamlit import config
-from streamlit.app_session import AppSession, AppSessionState
+from streamlit.runtime.app_session import AppSession, AppSessionState
 from streamlit.forward_msg_queue import ForwardMsgQueue
 from streamlit.proto.AppPage_pb2 import AppPage
 from streamlit.proto.BackMsg_pb2 import BackMsg
@@ -63,9 +63,9 @@ def _create_test_session(event_loop: Optional[AbstractEventLoop] = None) -> AppS
     )
 
 
-@patch("streamlit.app_session.LocalSourcesWatcher", MagicMock(spec=LocalSourcesWatcher))
+@patch("streamlit.runtime.app_session.LocalSourcesWatcher", MagicMock(spec=LocalSourcesWatcher))
 class AppSessionTest(unittest.TestCase):
-    @patch("streamlit.app_session.secrets._file_change_listener.disconnect")
+    @patch("streamlit.runtime.app_session.secrets._file_change_listener.disconnect")
     def test_shutdown(self, patched_disconnect):
         """Test that AppSession.shutdown behaves sanely."""
         session = _create_test_session()
@@ -126,14 +126,14 @@ class AppSessionTest(unittest.TestCase):
         clear_memo_cache.assert_called_once()
         clear_legacy_cache.assert_called_once()
 
-    @patch("streamlit.app_session.secrets._file_change_listener.connect")
+    @patch("streamlit.runtime.app_session.secrets._file_change_listener.connect")
     def test_request_rerun_on_secrets_file_change(self, patched_connect):
         """AppSession should add a secrets listener on creation."""
         session = _create_test_session()
         patched_connect.assert_called_once_with(session._on_secrets_file_changed)
 
     @patch_config_options({"runner.fastReruns": False})
-    @patch("streamlit.app_session.AppSession._create_scriptrunner")
+    @patch("streamlit.runtime.app_session.AppSession._create_scriptrunner")
     def test_rerun_with_no_scriptrunner(self, mock_create_scriptrunner: MagicMock):
         """If we don't have a ScriptRunner, a rerun request will result in
         one being created."""
@@ -142,7 +142,7 @@ class AppSessionTest(unittest.TestCase):
         mock_create_scriptrunner.assert_called_once_with(RerunData())
 
     @patch_config_options({"runner.fastReruns": False})
-    @patch("streamlit.app_session.AppSession._create_scriptrunner")
+    @patch("streamlit.runtime.app_session.AppSession._create_scriptrunner")
     def test_rerun_with_active_scriptrunner(self, mock_create_scriptrunner: MagicMock):
         """If we have an active ScriptRunner, it receives rerun requests."""
         session = _create_test_session()
@@ -160,7 +160,7 @@ class AppSessionTest(unittest.TestCase):
         mock_create_scriptrunner.assert_not_called()
 
     @patch_config_options({"runner.fastReruns": False})
-    @patch("streamlit.app_session.AppSession._create_scriptrunner")
+    @patch("streamlit.runtime.app_session.AppSession._create_scriptrunner")
     def test_rerun_with_stopped_scriptrunner(self, mock_create_scriptrunner: MagicMock):
         """If have a ScriptRunner but it's shutting down and cannot handle
         new rerun requests, we'll create a new ScriptRunner."""
@@ -179,7 +179,7 @@ class AppSessionTest(unittest.TestCase):
         mock_create_scriptrunner.assert_called_once_with(RerunData())
 
     @patch_config_options({"runner.fastReruns": True})
-    @patch("streamlit.app_session.AppSession._create_scriptrunner")
+    @patch("streamlit.runtime.app_session.AppSession._create_scriptrunner")
     def test_fast_rerun(self, mock_create_scriptrunner: MagicMock):
         """If runner.fastReruns is enabled, a rerun request will stop the
         existing ScriptRunner and immediately create a new one.
@@ -198,7 +198,7 @@ class AppSessionTest(unittest.TestCase):
         # And a new ScriptRunner should be created.
         mock_create_scriptrunner.assert_called_once()
 
-    @patch("streamlit.app_session.ScriptRunner")
+    @patch("streamlit.runtime.app_session.ScriptRunner")
     def test_create_scriptrunner(self, mock_scriptrunner: MagicMock):
         """Test that _create_scriptrunner does what it should."""
         session = _create_test_session()
@@ -226,8 +226,8 @@ class AppSessionTest(unittest.TestCase):
         )
         scriptrunner.start.assert_called_once()
 
-    @patch("streamlit.app_session.ScriptRunner", MagicMock(spec=ScriptRunner))
-    @patch("streamlit.app_session.AppSession._enqueue_forward_msg")
+    @patch("streamlit.runtime.app_session.ScriptRunner", MagicMock(spec=ScriptRunner))
+    @patch("streamlit.runtime.app_session.AppSession._enqueue_forward_msg")
     def test_ignore_events_from_noncurrent_scriptrunner(self, mock_enqueue: MagicMock):
         """If we receive ScriptRunnerEvents from anything other than our
         current ScriptRunner, we should silently ignore them.
@@ -261,7 +261,7 @@ class AppSessionTest(unittest.TestCase):
         session.request_rerun.assert_called_once_with(session._client_state)
 
     @patch(
-        "streamlit.app_session.AppSession._should_rerun_on_file_change",
+        "streamlit.runtime.app_session.AppSession._should_rerun_on_file_change",
         MagicMock(return_value=False),
     )
     def test_does_not_rerun_if_not_current_page(self):
@@ -273,7 +273,7 @@ class AppSessionTest(unittest.TestCase):
         self.assertEqual(session.request_rerun.called, False)
 
     @patch(
-        "streamlit.app_session.source_util.get_pages",
+        "streamlit.runtime.app_session.source_util.get_pages",
         MagicMock(
             return_value={
                 "hash1": {"page_name": "page1", "icon": "", "script_path": "script1"},
@@ -281,7 +281,7 @@ class AppSessionTest(unittest.TestCase):
             }
         ),
     )
-    @patch("streamlit.app_session.AppSession._enqueue_forward_msg")
+    @patch("streamlit.runtime.app_session.AppSession._enqueue_forward_msg")
     def test_on_pages_changed(self, mock_enqueue: MagicMock):
         session = _create_test_session()
         session._on_pages_changed("/foo/pages")
@@ -297,13 +297,13 @@ class AppSessionTest(unittest.TestCase):
         mock_enqueue.assert_called_once_with(expected_msg)
 
     @patch(
-        "streamlit.app_session.source_util.register_pages_changed_callback",
+        "streamlit.runtime.app_session.source_util.register_pages_changed_callback",
     )
     def test_installs_pages_watcher_on_init(self, patched_register_callback):
         session = _create_test_session()
         patched_register_callback.assert_called_once_with(session._on_pages_changed)
 
-    @patch("streamlit.app_session.source_util._on_pages_changed")
+    @patch("streamlit.runtime.app_session.source_util._on_pages_changed")
     def test_deregisters_pages_watcher_on_shutdown(self, patched_on_pages_changed):
         session = _create_test_session()
         session.shutdown()
@@ -341,11 +341,11 @@ class AppSessionScriptEventTest(IsolatedAsyncioTestCase):
     """Tests for AppSession's ScriptRunner event handling."""
 
     @patch(
-        "streamlit.app_session.config.get_options_for_section",
+        "streamlit.runtime.app_session.config.get_options_for_section",
         MagicMock(side_effect=_mock_get_options_for_section()),
     )
     @patch(
-        "streamlit.app_session.source_util.get_pages",
+        "streamlit.runtime.app_session.source_util.get_pages",
         MagicMock(
             return_value={
                 "hash1": {"page_name": "page1", "icon": "", "script_path": "script1"},
@@ -354,7 +354,7 @@ class AppSessionScriptEventTest(IsolatedAsyncioTestCase):
         ),
     )
     @patch(
-        "streamlit.app_session._generate_scriptrun_id",
+        "streamlit.runtime.app_session._generate_scriptrun_id",
         MagicMock(return_value="mock_scriptrun_id"),
     )
     async def test_enqueue_new_session_message(self):
@@ -450,11 +450,11 @@ class AppSessionScriptEventTest(IsolatedAsyncioTestCase):
         mock_handle_event.assert_called_once()
 
     @patch(
-        "streamlit.app_session.config.get_options_for_section",
+        "streamlit.runtime.app_session.config.get_options_for_section",
         MagicMock(side_effect=_mock_get_options_for_section()),
     )
     @patch(
-        "streamlit.app_session._generate_scriptrun_id",
+        "streamlit.runtime.app_session._generate_scriptrun_id",
         MagicMock(return_value="mock_scriptrun_id"),
     )
     async def test_handle_backmsg_exception(self):
@@ -543,7 +543,7 @@ class AppSessionScriptEventTest(IsolatedAsyncioTestCase):
 
 
 class PopulateCustomThemeMsgTest(unittest.TestCase):
-    @patch("streamlit.app_session.config")
+    @patch("streamlit.runtime.app_session.config")
     def test_no_custom_theme_prop_if_no_theme(self, patched_config):
         patched_config.get_options_for_section.side_effect = (
             _mock_get_options_for_section(
@@ -564,7 +564,7 @@ class PopulateCustomThemeMsgTest(unittest.TestCase):
 
         self.assertEqual(new_session_msg.HasField("custom_theme"), False)
 
-    @patch("streamlit.app_session.config")
+    @patch("streamlit.runtime.app_session.config")
     def test_can_specify_some_options(self, patched_config):
         patched_config.get_options_for_section.side_effect = _mock_get_options_for_section(
             {
@@ -585,7 +585,7 @@ class PopulateCustomThemeMsgTest(unittest.TestCase):
         # set to the type's zero value when undefined.
         self.assertEqual(new_session_msg.custom_theme.background_color, "")
 
-    @patch("streamlit.app_session.config")
+    @patch("streamlit.runtime.app_session.config")
     def test_can_specify_all_options(self, patched_config):
         patched_config.get_options_for_section.side_effect = (
             # Specifies all options by default.
@@ -600,8 +600,8 @@ class PopulateCustomThemeMsgTest(unittest.TestCase):
         self.assertEqual(new_session_msg.custom_theme.primary_color, "coral")
         self.assertEqual(new_session_msg.custom_theme.background_color, "white")
 
-    @patch("streamlit.app_session.LOGGER")
-    @patch("streamlit.app_session.config")
+    @patch("streamlit.runtime.app_session.LOGGER")
+    @patch("streamlit.runtime.app_session.config")
     def test_logs_warning_if_base_invalid(self, patched_config, patched_logger):
         patched_config.get_options_for_section.side_effect = (
             _mock_get_options_for_section({"base": "blah"})
@@ -616,8 +616,8 @@ class PopulateCustomThemeMsgTest(unittest.TestCase):
             " Allowed values include ['light', 'dark']. Setting theme.base to \"light\"."
         )
 
-    @patch("streamlit.app_session.LOGGER")
-    @patch("streamlit.app_session.config")
+    @patch("streamlit.runtime.app_session.LOGGER")
+    @patch("streamlit.runtime.app_session.config")
     def test_logs_warning_if_font_invalid(self, patched_config, patched_logger):
         patched_config.get_options_for_section.side_effect = (
             _mock_get_options_for_section({"font": "comic sans"})
@@ -634,7 +634,7 @@ class PopulateCustomThemeMsgTest(unittest.TestCase):
 
 
 @patch(
-    "streamlit.app_session.source_util.get_pages",
+    "streamlit.runtime.app_session.source_util.get_pages",
     MagicMock(
         return_value={
             "hash1": {"page_name": "page1", "script_path": "page1.py"},
