+++ lib/tests/streamlit/forward_msg_cache_test.py
@@ -18,7 +18,7 @@ from unittest.mock import MagicMock
 import unittest
 
 from streamlit import config, RootContainer
-from streamlit import app_session
+from streamlit.runtime import app_session
 from streamlit.forward_msg_cache import ForwardMsgCache
 from streamlit.forward_msg_cache import create_reference_msg
 from streamlit.forward_msg_cache import populate_hash_if_needed
