+++ lib/tests/streamlit/web/server/server_test_case.py
@@ -21,7 +21,7 @@ import tornado.web
 import tornado.websocket
 from tornado.websocket import WebSocketClientConnection
 
-from streamlit.app_session import AppSession
+from streamlit.runtime.app_session import AppSession
 from streamlit.proto.ForwardMsg_pb2 import ForwardMsg
 from streamlit.web.server import Server
